+++ src/app/Balloon.App.Cli/Console/Jobs.php
@@ -15,6 +15,7 @@ use Balloon\Hook;
 use GetOpt\GetOpt;
 use Psr\Log\LoggerInterface;
 use TaskScheduler\Queue;
+use TaskScheduler\Scheduler;
 
 class Jobs
 {
@@ -40,6 +41,13 @@ class Jobs
     protected $queue;
 
     /**
+     * Scheduler.
+     *
+     * @var Scheduler
+     */
+    protected $scheduler;
+
+    /**
      * Hook.
      *
      * @var Hook
@@ -49,13 +57,13 @@ class Jobs
     /**
      * Constructor.
      */
-    public function __construct(Hook $hook, Queue $queue, LoggerInterface $logger, GetOpt $getopt)
+    public function __construct(Hook $hook, Queue $queue, Scheduler $scheduler, LoggerInterface $logger, GetOpt $getopt)
     {
         $this->queue = $queue;
         $this->hook = $hook;
         $this->logger = $logger;
         $this->getopt = $getopt;
-        $this->queue = $queue;
+        $this->scheduler = $scheduler;
     }
 
     /**
@@ -67,13 +75,19 @@ class Jobs
             'category' => get_class($this),
         ]);
 
+        if ($this->getopt->getOption('flush')) {
+            $this->scheduler->flush();
+        }
+
         $this->hook->run('preExecuteAsyncJobs');
         $this->queue->process();
 
         return true;
     }
 
-    // Get operands
+    /**
+     * Get operands.
+     */
     public static function getOperands(): array
     {
         return [];
@@ -84,6 +98,8 @@ class Jobs
      */
     public static function getOptions(): array
     {
-        return [];
+        return [
+            \GetOpt\Option::create('f', 'flush')->setDescription('Flush queue before start (Attention all jobs get removed)'),
+        ];
     }
 }