+++ src/lib/Async/WorkerFactory.php
@@ -15,11 +15,10 @@ use Balloon\Bootstrap\ContainerBuilder;
 use Composer\Autoload\ClassLoader as Composer;
 use ErrorException;
 use MongoDB\BSON\ObjectId;
-use MongoDB\Database;
 use Psr\Log\LoggerInterface;
-use TaskScheduler\Scheduler;
 use TaskScheduler\Worker;
 use TaskScheduler\WorkerFactoryInterface;
+use TaskScheduler\WorkerManager;
 
 class WorkerFactory implements WorkerFactoryInterface
 {
@@ -41,12 +40,25 @@ class WorkerFactory implements WorkerFactoryInterface
     /**
      * {@inheritdoc}
      */
-    public function build(ObjectId $id): Worker
+    public function buildWorker(ObjectId $id): Worker
     {
         $dic = ContainerBuilder::get($this->composer);
         $this->setErrorHandler($dic->get(LoggerInterface::class));
 
-        return new Worker($id, $dic->get(Scheduler::class), $dic->get(Database::class), $dic->get(LoggerInterface::class), $dic);
+        return $dic->make(Worker::class, [
+            'id' => $id,
+        ]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function buildManager(): WorkerManager
+    {
+        $dic = ContainerBuilder::get($this->composer);
+        $this->setErrorHandler($dic->get(LoggerInterface::class));
+
+        return $dic->get(WorkerManager::class);
     }
 
     /**