+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## 2.4.0
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
-**Date**: Mon Jan 28 14:34:01 CET 2019\
+**Date**: Tue Jan 29 14:42:01 CET 2019\
 **Note**: queue flush required, @see UPGRADE.md
 
 * CORE: [CHANGE] Upgrade to stable v3 \TaskScheduler #220, possibilty to flush job queue using ballooncli jobs -f
-* CORE: [CHANGE] Create a new file should check if there is an existing , deleted one with the same name under the same collection #241
+* CORE: [CHANGE] Create a new file should check if there is an existing, deleted one with the same name under the same collection #241
 * CORE: [FIX] Missing copy delta entries #252
 * API: [FIX] Call to undefined method Balloon\Filesystem\Node\Collection::put() #249
 * API: [FIX] Missing nodes in delta if delta requested for single node #251
@@ -58,7 +58,6 @@
 **Date**: Wed Dez 05 08:50:32 CET 2018\
 
 * API: [FEATURE] Progressive stream via http #199 
-* CORE: [CHANGE] Upgrade to stable v3 \TaskScheduler #220, possibilty to flush job queue using ballooncli jobs -f
 
 
 ## 2.1.9